+++ invocations/checks.py
@@ -10,7 +10,9 @@ from invoke import task
 
 
 @task(name="blacken", iterable=["folder"])
-def blacken(c, line_length=79, folder=None, check=False, diff=False):
+def blacken(
+    c, line_length=79, folder=None, check=False, diff=False, find_opts=None
+):
     """
     Run black on the current source tree (all ``.py`` files).
 
@@ -28,8 +30,14 @@ def blacken(c, line_length=79, folder=None, check=False, diff=False):
         Whether to run ``black --check``. Default: ``False``.
     :param bool diff:
         Whether to run ``black --diff``. Default: ``False``.
+    :param str find_opts:
+        Extra option string appended to the end of the internal ``find``
+        command. For example, skip a vendor directory with ``"-and -not -path
+        ./vendor\*"``, add ``-mtime N``, or etc.
 
     .. versionadded:: 1.2
+    .. versionchanged:: 1.4
+        Added the ``find_opts`` argument.
     """
     default_folders = ["."]
     configured_folders = c.config.get("blacken", {}).get(
@@ -42,8 +50,12 @@ def blacken(c, line_length=79, folder=None, check=False, diff=False):
         black_command_line = "{} --check".format(black_command_line)
     if diff:
         black_command_line = "{} --diff".format(black_command_line)
+    if find_opts is not None:
+        find_opts = " {}".format(find_opts)
+    else:
+        find_opts = ""
 
-    cmd = "find {} -name '*.py' | xargs {}".format(
-        " ".join(folders), black_command_line
+    cmd = "find {} -name '*.py'{} | xargs {}".format(
+        " ".join(folders), find_opts, black_command_line
     )
     c.run(cmd, pty=True)