+++ docs/changelog.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+- :feature:`-` Add a ``find_opts`` argument to ``checks.blacken`` for improved
+  control over what files get blackened.
 - :release:`1.3.0 <2018-06-20>`
 - :feature:`-` Bump Releases requirement up to 1.6 and leverage its new ability
   to load Sphinx extensions, in ``packaging.release.prepare`` (which parses